+++ experiments/scripts/A2_extract_topk.py
@@ -2,10 +2,45 @@ import sys
 import numpy as np
 import pandas as pd
 from pathlib import Path
-from utils.metrics import compute_diverse_top_k
+from rdkit import Chem, DataStructs
+from rdkit import RDLogger
+
+RDLogger.DisableLog("rdApp.*")
+
+
+def compute_diverse_top_k(
+    smiles: list[str],
+    rewards: list[float],
+    k: int,
+    thresh: float = 0.5,
+) -> list[tuple[int, str, float]]:
+    modes = [(i, smi, r) for i, (r, smi) in enumerate(zip(rewards, smiles, strict=True))]
+    modes.sort(key=lambda m: m[2], reverse=True)
+    top_modes = [modes[0]]
+
+    prev_smis = {modes[0][1]}
+    mode_fps = [Chem.RDKFingerprint(Chem.MolFromSmiles(modes[0][1]))]
+    for i in range(1, len(modes)):
+        smi = modes[i][1]
+        if smi in prev_smis:
+            continue
+        prev_smis.add(smi)
+        if thresh > 0:
+            fp = Chem.RDKFingerprint(Chem.MolFromSmiles(smi))
+            sim = DataStructs.BulkTanimotoSimilarity(fp, mode_fps)
+            if max(sim) >= (1 - thresh):  # div = 1- sim
+                continue
+            mode_fps.append(fp)
+            top_modes.append(modes[i])
+        else:
+            top_modes.append(modes[i])
+        if len(top_modes) >= k:
+            break
+    return top_modes
+
 
 ROOT_DIR = Path(sys.argv[1])
-SAVE_DIR = Path("./result/benchmark/") / ROOT_DIR.name / "topk"
+SAVE_DIR = Path("./result/benchmark/topk/") / ROOT_DIR.name
 SAVE_DIR.mkdir(parents=True, exist_ok=True)
 
 TARGETS = [
@@ -27,9 +62,9 @@ TARGETS = [
 ]
 
 avg_scores = []
-for target in TARGETS[:10]:
+for target in TARGETS:
     scores = []
-    for trial in [0]:
+    for trial in [0, 1, 2]:
         save_dir = SAVE_DIR / target
         save_dir.mkdir(exist_ok=True)
         save_filename = save_dir / f"seed-{trial}.csv"
